// src/app/shared/interfaces/person.ts
// export interface Address {
//    area: string;
//    road: string;
// }

export interface Person {
   givenName: string;
   surName: string;
   age: number;
   email: string;
   address: string;
   // address: Address;    εναλακτικα θα μπορουσα να κανω και αυτο
}